+++ drivers/timer/nrf_rtc_timer.c
@@ -8,7 +8,7 @@
 #include <soc.h>
 #include <clock_control.h>
 #include <drivers/clock_control/nrf_clock_control.h>
-#include <system_timer.h>
+#include <drivers/timer/system_timer.h>
 #include <sys_clock.h>
 #include <nrf_rtc.h>
 #include <spinlock.h>
