+++ drivers/timer/rv32m1_lptmr_timer.c
@@ -6,7 +6,7 @@
 
 #include <zephyr.h>
 #include <misc/util.h>
-#include <system_timer.h>
+#include <drivers/timer/system_timer.h>
 #include <soc.h>
 
 /*
