+++ drivers/timer/sam0_rtc_timer.c
@@ -16,7 +16,7 @@
 
 #include <soc.h>
 #include <clock_control.h>
-#include <system_timer.h>
+#include <drivers/timer/system_timer.h>
 #include <sys_clock.h>
 
 /* RTC registers. */
