+++ kernel/idle.c
@@ -8,7 +8,7 @@
 #include <kernel_structs.h>
 #include <toolchain.h>
 #include <linker/sections.h>
-#include <drivers/system_timer.h>
+#include <drivers/timer/system_timer.h>
 #include <wait_q.h>
 #include <power.h>
 #include <stdbool.h>
